boolean filter + add trait
