+++ src/PowerGridComponent.php
@@ -20,8 +20,8 @@ use PowerComponents\LivewirePowerGrid\Traits\{BatchableExport,
     Filter,
     Listeners,
     PersistData,
-    WithSorting
-};
+    WithDynamicFilters,
+    WithSorting};
 use Throwable;
 
 class PowerGridComponent extends Component
@@ -35,6 +35,7 @@ class PowerGridComponent extends Component
     use BatchableExport;
     use PersistData;
     use Listeners;
+    use WithDynamicFilters;
 
     public array $headers = [];
 
@@ -141,41 +142,8 @@ class PowerGridComponent extends Component
         });
     }
 
-    private function initializePropertiesFromDynamicFilters(): void
-    {
-        $this->columns = collect($this->columns)->transform(function (Column $column) {
-            if (Support\Arr::has($column->inputs, 'dynamic')) {
-                $filterType = strval(data_get($column->inputs, 'dynamic.filterType'));
-                $dataField  = strval(data_get($column->inputs, 'dynamic.dataField'));
-
-                if (!in_array($filterType, DynamicInput::DYNAMIC_FILTERS)) {
-                    throw new Exception('Available options must be included by: \PowerComponents\LivewirePowerGrid\DynamicInput::class');
-                }
-
-                $filter     = strval(Support\Str::of($filterType)->explode(',')->get(0));
-                $type       = strval(Support\Str::of($filterType)->explode(',')->get(1));
-
-                /** @phpstan-ignore-next-line */
-                $initial = match ($type) {
-                    'array'   => [],
-                    'string'  => '',
-                    'boolean' => null,
-                };
-
-                $this->filters[$filter][$dataField] = $initial;
-
-                data_set($column->inputs, 'dynamic.attributes', [
-                    ...(array) data_get($column->inputs, 'dynamic.attributes'),
-                    'wire:model.debounce.500ms' => 'filters.' . $filter . '.' . $dataField,
-                ]);
-            }
-
-            return $column;
-        })->toArray();
-    }
-
     /**
-     * @throws Exception
+     * @throws Exception | Throwable
      */
     public function render(): Application|Factory|View
     {
